requirements.txt
autogen==0.2.*
gradio==4.0.*
pydantic==2.0.*
pydantic-settings==2.0.*
openai==1.0.*
numpy==1.24.*
pillow==10.0.*
requests==2.31.*
pyyaml==6.0.*

nltk==3.8.*
scikit-learn==1.3.*

stability-sdk==0.8.*

transformers==4.36.*
diffusers==0.24.*
accelerate==0.25.*

matplotlib==3.7.*
plotly==5.16.*

pytest==7.4.*
black==23.9.*
flake8==6.1.*